python-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.0.350
openai==0.28.1
pydantic==1.10.13
chromadb==0.4.22
tiktoken==0.5.2